requirements.txt
jinja2
pyyaml
mkdocs
mkdocs-material
mkdocs-awesome-pages-plugin
openai
jsonschema
fastapi
uvicorn
requests
redis
structlog
